+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ### 1.0.3
++ [x] Events: add `$(document).ready(...)`
 
 ### 1.0.2
 + [x] Events: fix `function.name` property for IE11